dsp: factor out tools
